pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "lifelines>=0.30.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "scikit-survival>=0.24.1",
    "scipy>=1.15.2",
    "streamlit>=1.44.0",
    "tensorflow>=2.14.0",
]
